+++ src/ska_low_mccs_spshw/tile/attribute_managers.py
@@ -7,6 +7,7 @@
 """A collection of AttributeManagers to serve the TANGO attribute."""
 from __future__ import annotations
 
+import math
 import time
 from typing import Any, Callable
 
@@ -74,13 +75,21 @@ class AttributeManager:
         :param value: the value we want to update attribute with.
         :param post: Optional flag to post an update.
         """
+
+        def _is_none_type(val: Any) -> bool:
+            if val is None or (isinstance(val, float) and math.isnan(val)):
+                return True
+            return False
+
         is_event_required = False
         self._last_update = time.time()
-        new_value: Any = (
-            self._converter(value) if self._converter and value is not None else value
+        new_value = (
+            self._converter(value)
+            if self._converter and not _is_none_type(value)
+            else value
         )
 
-        if new_value is None:
+        if _is_none_type(new_value):
             # Move to invalid and require event!
             if self._quality != tango.AttrQuality.ATTR_INVALID:
                 self._quality = tango.AttrQuality.ATTR_INVALID