+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ## 10.0.1
 
 * [THORN-313] Loop over antennas present in mapping in station_component_manager._calculate_delays_per_tile instead of looping over delays. Improved some messages/docstrings.
+* [THORN-261] Added property UseAttributesForHealth. A new feature toggle activating ADR-115 in MccsTile. By default this is True, when True the attribute
+quality factor will be used to inform health.
+* [THORN-261] bump ska-low-sps-tpm-api 1.0.0 -> 1.0.1 (<https://gitlab.com/ska-telescope/ska-low-sps-tpm/ska-low-sps-tpm-api/-/releases/1.0.1>)
 
 ## 10.0.0
 
@@ -125,6 +128,7 @@
 * [THORN-249] Add attribute pfb_version to MccsTile.
 * [THORN-249] Bump firmware version 6.6.1 -> 6.7.1
 * [SKB-1035] Added tile programming state to the health rollup
+* [THORN-311] Added target_adc and bias parameters to TriggerAdcEqualisation
 
 ## 9.0.0